% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_loss.R
\name{calculate_loss}
\alias{calculate_loss}
\title{Calculate pairwise Loss fnuctoins}
\usage{
calculate_loss(
  predictions,
  loss_fns = NULL,
  na_value = NaN,
  task,
  resampling = NULL,
  iters = seq_along(predictions)
)
}
\arguments{
\item{predictions}{(list of \code{\link{Prediction}}s)\cr
A list containing the predictions.}

\item{loss_fns}{(named \code{list()} of loss \code{function}s)\cr
The (pairwise) loss functions that are to be calculated.
The names of the functions will be the names in the returned loss table.}
}
\value{
\code{\link{data.table}}
}
\description{
Evaluate one or more pairwise loss functions on a list of predictions.
}
