% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendCached.R
\name{DataBackendCached}
\alias{DataBackendCached}
\title{Cached Data Backend}
\description{
This backend wraps another backend -- usually an out of memory backend -- and when first accessing
the data field, it creates a \code{\link[mlr3:DataBackendDataTable]{mlr3::DataBackendDataTable}} for the IDs provided
during construction.
}
\section{Super class}{
\code{\link[mlr3:DataBackend]{mlr3::DataBackend}} -> \code{DataBackendCached}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataBackendCached-new}{\code{DataBackendCached$new()}}
\item \href{#method-DataBackendCached-data}{\code{DataBackendCached$data()}}
\item \href{#method-DataBackendCached-missings}{\code{DataBackendCached$missings()}}
\item \href{#method-DataBackendCached-distinct}{\code{DataBackendCached$distinct()}}
\item \href{#method-DataBackendCached-head}{\code{DataBackendCached$head()}}
\item \href{#method-DataBackendCached-clone}{\code{DataBackendCached$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="DataBackend" data-id="format"><a href='../../mlr3/html/DataBackend.html#method-DataBackend-format'><code>mlr3::DataBackend$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="DataBackend" data-id="print"><a href='../../mlr3/html/DataBackend.html#method-DataBackend-print'><code>mlr3::DataBackend$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBackendCached-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataBackendCached-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendCached$new(backend, ids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBackendCached-data"></a>}}
\if{latex}{\out{\hypertarget{method-DataBackendCached-data}{}}}
\subsection{Method \code{data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendCached$data(rows, cols, data_format = "data.table")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBackendCached-missings"></a>}}
\if{latex}{\out{\hypertarget{method-DataBackendCached-missings}{}}}
\subsection{Method \code{missings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendCached$missings(rows, cols)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBackendCached-distinct"></a>}}
\if{latex}{\out{\hypertarget{method-DataBackendCached-distinct}{}}}
\subsection{Method \code{distinct()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendCached$distinct(rows, cols, na_rm = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBackendCached-head"></a>}}
\if{latex}{\out{\hypertarget{method-DataBackendCached-head}{}}}
\subsection{Method \code{head()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendCached$head(n = 6L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataBackendCached-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataBackendCached-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendCached$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
