% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_52cv.R
\name{infer_52cv}
\alias{infer_52cv}
\title{Paired T-Test with 5x2 Cross-Validation}
\usage{
infer_52cv(x, alpha = 0.05)
}
\arguments{
\item{x}{(\code{\link{ResampleResult}} or  \code{\link{BenchmarkResult}})\cr
The resample result or benchmark result using a \code{\link{ResamplingNestedCV}} as the resampling technique.}

\item{alpha}{(\code{numeric(1)})\cr
The alpha level for the confidence interval.}

\item{loss}{(\code{character(1)})\cr
The observation loss. One of \code{mlr3measures::measures} that calculates oberservation-wise losses in an unaggergated
manner. E.g. \code{"se"} for the squared error or \code{"zero_one"} for the 0-1 loss.}
}
\description{
Approximate t-test using 5x2 Cross-Validation.
}
\references{
Dietterich, G T (1998).
\dQuote{Approximate statistical tests for comparing supervised classification learning algorithms.}
\emph{Neural computation}, \bold{10}(7), 1895--1923.
}
