% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_bates.R
\name{infer_bates}
\alias{infer_bates}
\title{Inference Method for Nested Cross-Validation}
\usage{
infer_bates(x, alpha, loss, ...)
}
\arguments{
\item{x}{(\code{\link{ResampleResult}} or  \link{``})\cr
The resample result or benchmark result using a \code{\link{ResamplingNestedCV}} as the resampling technique.}

\item{alpha}{(\code{numeric(1)})\cr
The alpha level for the confidence interval.}

\item{loss}{(\code{character(1)})\cr
The observation loss. One of \code{mlr3measures::measures} that calculates oberservation-wise losses in an unaggergated
manner. E.g. \code{"se"} for the squared error or \code{"zero_one"} for the 0-1 loss.}
}
\description{
Inference method for Nested Cross-Validation.
}
\references{
Bates, Stephen, Hastie, Trevor, Tibshirani, Robert (2021).
\dQuote{Cross-validation: what does it estimate and how well does it do it?}
\doi{10.48550/ARXIV.2104.00673}, \url{https://arxiv.org/abs/2104.00673}.
}
