% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_conservative_z.R
\name{infer_conservative_z}
\alias{infer_conservative_z}
\title{Conservative Z-Test}
\usage{
infer_conservative_z(x, alpha = 0.05)
}
\arguments{
\item{x}{(\code{\link{ResampleResult}} or  \code{\link{BenchmarkResult}})\cr
The resample result or benchmark result}
}
\description{
Conservative Z-Test.
}
