% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_corrected_t.R
\name{infer_corrected_t}
\alias{infer_corrected_t}
\title{Corrected T-Test}
\usage{
infer_corrected_t(x, alpha = 0.05, loss)
}
\arguments{
\item{x}{(\code{\link{ResampleResult}} or  \code{\link{BenchmarkResult}})\cr
The resampling must be a \code{\link{ResamplingSubsampling}} with more than one repetition (parameter \code{repeats}).}

\item{alpha}{(\code{numeric(1)})\cr
The alpha level for the confidence interval.}

\item{loss}{}
}
\description{
Applies the corrected T-Test.
}
\references{
Nadeau, Claude, Bengio, Yoshua (1999).
\dQuote{Inference for the generalization error.}
\emph{Advances in neural information processing systems}, \bold{12}.
}
