% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_naive_cv.R
\name{infer_naive_cv}
\alias{infer_naive_cv}
\title{Naive CV Estimation for}
\usage{
infer_naive_cv(x, alpha, loss, within_folds = TRUE)
}
\arguments{
\item{x}{(any)\cr
Whether to}

\item{loss}{}
}
\description{
Naive CV Estimation for
}
