% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingBootstrapCCV.R
\name{mlr_resamplings_bootstrap_ccv}
\alias{mlr_resamplings_bootstrap_ccv}
\alias{ResamplingBootstrapCCV}
\title{Bootstrap Case Cross-Validation Percentile findInterval()}
\description{
Create bootstrap samples like in \code{\link[mlr3:mlr_resamplings_bootstrap]{mlr3::ResamplingBootstrap}}.
For each of these bootstrap samples, conduct "leave one case out" (LOCO) cross-validation.
This implies that the training sizes in the LOCO differ.
}
\section{Parameters}{

\itemize{
\item \code{repeats} :: (\code{integer(1)})\cr
Number of repetitions.
\item \code{ratio} :: (\code{numeric(1)})\cr
Ratio of observations to put into the training set.
}
}

\section{Dictionary}{

This \link{Resampling} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_resamplings} or with the associated sugar function \code{\link[=rsmp]{rsmp()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_resamplings$get("bootstrap_ccv")
rsmp("bootstrap_ccv")
}\if{html}{\out{</div>}}
}

\examples{
res = rsmp("bootstrap_ccv", ratio = 0.5, repeats = 2)
task = tsk("penguins")
res$instantiate(task)
res
}
\references{
Jiang, Wenyu, Varma, Sudhir, Simon, Richard (2008).
\dQuote{Calculating Confidence Intervals for Prediction Error in Microarray Classification Using Resampling.}
\emph{Statistical applications in genetics and molecular biology}, \bold{7}, Article8.
\doi{10.2202/1544-6115.1322}.
}
\section{Super class}{
\code{\link[mlr3:Resampling]{mlr3::Resampling}} -> \code{ResamplingBootstrapCCV}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{iters}}{(\code{integer(1)})\cr
The number of iterations.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResamplingBootstrapCCV-new}{\code{ResamplingBootstrapCCV$new()}}
\item \href{#method-ResamplingBootstrapCCV-clone}{\code{ResamplingBootstrapCCV$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="format"><a href='../../mlr3/html/Resampling.html#method-Resampling-format'><code>mlr3::Resampling$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="help"><a href='../../mlr3/html/Resampling.html#method-Resampling-help'><code>mlr3::Resampling$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="instantiate"><a href='../../mlr3/html/Resampling.html#method-Resampling-instantiate'><code>mlr3::Resampling$instantiate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="print"><a href='../../mlr3/html/Resampling.html#method-Resampling-print'><code>mlr3::Resampling$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="test_set"><a href='../../mlr3/html/Resampling.html#method-Resampling-test_set'><code>mlr3::Resampling$test_set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="train_set"><a href='../../mlr3/html/Resampling.html#method-Resampling-train_set'><code>mlr3::Resampling$train_set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingBootstrapCCV-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingBootstrapCCV-new}{}}}
\subsection{Method \code{new()}}{
creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingBootstrapCCV$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingBootstrapCCV-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingBootstrapCCV-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingBootstrapCCV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
