% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskGeneratorBates.R
\name{mlr_task_generators_bates_regr}
\alias{mlr_task_generators_bates_regr}
\alias{TaskGeneratorBatesClassif}
\title{Bates Classification Task Generator}
\description{
A \link[mlr3:TaskGenerator]{mlr3::TaskGenerator} for the bates task.
}
\section{Super class}{
\code{\link[mlr3:TaskGenerator]{mlr3::TaskGenerator}} -> \code{TaskGeneratorBatesClassif}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaskGeneratorBatesClassif-new}{\code{TaskGeneratorBatesClassif$new()}}
\item \href{#method-TaskGeneratorBatesClassif-clone}{\code{TaskGeneratorBatesClassif$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="TaskGenerator" data-id="format"><a href='../../mlr3/html/TaskGenerator.html#method-TaskGenerator-format'><code>mlr3::TaskGenerator$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="TaskGenerator" data-id="generate"><a href='../../mlr3/html/TaskGenerator.html#method-TaskGenerator-generate'><code>mlr3::TaskGenerator$generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="TaskGenerator" data-id="print"><a href='../../mlr3/html/TaskGenerator.html#method-TaskGenerator-print'><code>mlr3::TaskGenerator$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGeneratorBatesClassif-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGeneratorBatesClassif-new}{}}}
\subsection{Method \code{new()}}{
creates a new instance of this \link[r6:r6class]{r6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGeneratorBatesClassif$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGeneratorBatesClassif-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGeneratorBatesClassif-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGeneratorBatesClassif$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
