% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingNestedCV.R
\name{ResamplingNestedCV}
\alias{ResamplingNestedCV}
\title{Nested Cross-Validation}
\description{
We have R times:
\itemize{
\item K outer resamplings
\item K * (K - 1) inner resamplings
}

The first K^2 iters are then:
The first K are the outer K iterations.
Then, we have K * (K - 1) inner iterations.
}
\section{Super class}{
\code{\link[mlr3:Resampling]{mlr3::Resampling}} -> \code{ResamplingNestedCV}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{iters}}{(\code{integer(1)})\cr
The number of iterations.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResamplingNestedCV-new}{\code{ResamplingNestedCV$new()}}
\item \href{#method-ResamplingNestedCV-unflatten}{\code{ResamplingNestedCV$unflatten()}}
\item \href{#method-ResamplingNestedCV-flatten}{\code{ResamplingNestedCV$flatten()}}
\item \href{#method-ResamplingNestedCV-clone}{\code{ResamplingNestedCV$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="format"><a href='../../mlr3/html/Resampling.html#method-Resampling-format'><code>mlr3::Resampling$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="help"><a href='../../mlr3/html/Resampling.html#method-Resampling-help'><code>mlr3::Resampling$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="instantiate"><a href='../../mlr3/html/Resampling.html#method-Resampling-instantiate'><code>mlr3::Resampling$instantiate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="print"><a href='../../mlr3/html/Resampling.html#method-Resampling-print'><code>mlr3::Resampling$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="test_set"><a href='../../mlr3/html/Resampling.html#method-Resampling-test_set'><code>mlr3::Resampling$test_set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="train_set"><a href='../../mlr3/html/Resampling.html#method-Resampling-train_set'><code>mlr3::Resampling$train_set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingNestedCV-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingNestedCV-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingNestedCV$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingNestedCV-unflatten"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingNestedCV-unflatten}{}}}
\subsection{Method \code{unflatten()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingNestedCV$unflatten(iter)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingNestedCV-flatten"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingNestedCV-flatten}{}}}
\subsection{Method \code{flatten()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingNestedCV$flatten(rep, outer, inner = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResamplingNestedCV-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResamplingNestedCV-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingNestedCV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
